% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{AdverseEvents}
\alias{AdverseEvents}
\title{AdverseEvents}
\usage{
AdverseEvents(event_filename)
}
\arguments{
\item{event_filename}{The name of the flat file containing the event specific information extracted from MEDLINE}
}
\value{
An object containing for adverse event candidates: names, MeSH codes, and MEDLINE Ids for papers mentioning potentially adverse events
}
\description{
This function creates a data table with the event candidates extracted from MEDLINE
}
\examples{
{

   events <- AdverseEvents("manifestations_MEDLINE_360k.txt")
   
   #will return the PubMed IDs for the 933 papers in MEDLINE mentioning Epilepsy as an adverse event
   events[event=="Epilepsy",]$PMID
    
} 

}

