% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{Drugs}
\alias{Drugs}
\title{Drugs}
\usage{
Drugs(drug_filename, classLevel = FALSE)
}
\arguments{
\item{drug_filename}{The name of the flat file containing the drug specific information extracted from MEDLINE}

\item{classLevel}{If TRUE, represent drugs at the drug class level, e.g. "bupivacaine" as "amides"}
}
\value{
An object containing for drug candidates: names, ATC codes, and MEDLINE Ids for papers mentioning the drugs potentially in the context of adverse events
}
\description{
This function creates a data table with the drug candidates extracted from MEDLINE
}
\examples{
{

   drugs <- Drugs("drugs_ATC_MEDLINE_360k.txt")
   
   #will return the PubMed IDs for the 758 papers in MEDLINE mentioning the drug bupivacaine together with an adverse event
   drugs[name=="bupivacaine",]$PMID
    
} 

}

