% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getAbstractionLayer}
\alias{getAbstractionLayer}
\title{Abstraction Layer on annotation terms (MeSH)}
\usage{
getAbstractionLayer(ATC_annotations, MeSH_annotations, range = c(),
  fix = c())
}
\arguments{
\item{ATC_annotations}{The ATC code for the drug of interest.}

\item{MeSH_annotations}{The MeSH term for the adverse event of interest.}

\item{range}{The abstraction layer range based on Information Content of terms to be included.}

\item{fixed}{The abstraction level is defined by a fixed level in the MeSH tree hierarchcy (e.g. level 2).}
}
\value{
An object containing the abstraction level information for the specified level (list containing the range, the layer, the drugs used as the input parameter, the terms in the annotation file that are covered at the specified abstraction layer, and a set of higher-level terms that represent these terms).
}
\description{
This function creates an Abstraction Layer object that can be used to represent input data (e.g., adverse event terms) and a background set (all MEDLINE) at a given abstraction level.
}
\details{
This function uses hierarchical information from MeSH and Information Content (IC) that has been precalculated for every MeSH term based on aggregated counts on the entire MEDLINE background set to aggregate individual specific terms into more general terms from the specified abstraction level. Alternatively, MeSH terms can be mapped to parent terms at a specific fixed level in the MeSH tree hierarchy (e.g., 2nd level).
}
\examples{
{

  library("GEA")
  
  drugs <- Drugs("drugs_ATC_MEDLINE_360k.txt")
  events <- AdverseEvents("manifestations_MEDLINE_360k.txt")
  abs710    <- getAbstractionLayer(drugs,events,c(7,10))
}

}

