% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getEnrichment}
\alias{getEnrichment}
\title{This function executes the enrichment analysis}
\usage{
getEnrichment(set_of_interest, abstraction = getAbstractionLayer(c(4, 7)),
  threshold = 0.1, filter = c(), adjust = FALSE)
}
\arguments{
\item{set_of_interest}{The set of articles for a given set of interest (e.g., a drug, several drugs, a drug class).}

\item{abstraction}{The abstraction layer object for specified abstraction level.}

\item{threshold}{Only terms with p-values below threshold are returned.}

\item{filter}{Instead for all terms in the set of interest only for the specified terms enrichment will be calculated.}

\item{adjust}{If set to TRUE, conditional pvalues will be calculated that adjust for known dependencies among terms.}
}
\value{
Table of enriched terms for given set of interest ordered by p-value.
}
\description{
This function will analyze which terms are enriched for a given set of interest at the given abstraction level. A list of all terms in the set will be returned together with counts and p-values for enrichment.
}
\details{
Intersects set of interest with MeSH annotation terms, aggregates terms to specified abstraction level, calculates enrichment for these terms against the general background set (all MEDLINE). All terms that are enrichend (below a given threshold, default = 0.1) will be returned. If filter is set, from all enriched terms only those in the filter set are returned. If the adjust flag is set, enriched terms will be checked for known dependencies based on pairwise conditional hypergeometric test among all enriched terms. This might take considerably longer than the unadjusted EA.
}
\examples{
{

  library("GEA")
  
  drugs <- Drugs("drugs_ATC_MEDLINE_360k.txt")
  events <- AdverseEvents("manifestations_MEDLINE_360k.txt")
  abs710    <- getAbstractionLayer(drugs,events,c(7,10))
  
  getEnrichment(drugs[ code== "C01EB03" ,] , abs710)$annotation_profile
}

}

