% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPRR}
\alias{getPRR}
\title{This function calculates the Proportional Reporting Ratio (PRR) for a drug and an adverse event.}
\usage{
getPRR(atccode, event, abstraction = getAbstractionLayer(c(4, 7)))
}
\arguments{
\item{atccode}{The ATC code for the drug of interest.}

\item{event}{The MeSH term for the adverse event of interest.}

\item{abstraction}{The abstraction layer object for specified abstraction level.}
}
\value{
PRR for given drug - adverse event pair
}
\description{
This function will detect signals for given adverse drug event pair based on the biomedical literature (MEDLINE).
}
\details{
Constructs a contingency table for given drug and adverse event based on theit occurrrences and co-occurrences in the literature set (provided by the abstraction layer object), aggregates terms to specified abstraction level. Performs zero-cell correction. Returns PRR for given drug - adverse event pair.
}
\examples{
{

  library("GEA")
  
  drugs <- Drugs("drugs_ATC_MEDLINE_360k.txt")
  events <- AdverseEvents("manifestations_MEDLINE_360k.txt")
  abs710    <- getAbstractionLayer(drugs,events,c(7,10))
  
  getPRR("R03DC01","D058186",abs710)
}

}

