% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPValue}
\alias{getPValue}
\title{This function calculates the p-value for GEA.}
\usage{
getPValue(drug, event, abstraction)
}
\arguments{
\item{event}{The MeSH term for the adverse event of interest.}

\item{abstraction}{The abstraction layer object for specified abstraction level.}

\item{atccode}{The ATC code for the drug of interest.}
}
\value{
P-value for co-occurrence of a given drug - adverse event pair.
}
\description{
This function will detect signals for a given adverse drug event pair based on the biomedical literature (MEDLINE) using GEA.
}
\details{
Calculates significane of co-occurrence of drug and adverse event in comparison to large reference set (all MEDLINE). Returns p-value for given drug - adverse event pair.
}
\examples{
{

  library("GEA")
  
  drugs <- Drugs("drugs_ATC_MEDLINE_360k.txt")
  events <- AdverseEvents("manifestations_MEDLINE_360k.txt")
  abs710    <- getAbstractionLayer(drugs,events,c(7,10))
  
  getPValue("R03DC01","D058186",abs710)
}

}

