% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{GS}
\alias{GS}
\title{GS}
\usage{
GS(gs_name)
}
\arguments{
\item{gs_name}{The name of the gold standard file}
}
\value{
An object containing the Gold Standard data.
}
\description{
This function creates an Gold Standard object that can be used to evaluate the performance of drug adverse event signal detection methods.
}
\examples{
{

   gs <- GS("OMOP_GS_drugs.csv")

} 

}

